Extract common functionality into helpers
