+++ test/test_request.py
@@ -1,27 +1,7 @@
 import unittest
-import asyncio
-import io
+from helpers import FakeReader, a_wait
 
 import grole
-
-a_wait = asyncio.new_event_loop().run_until_complete
-
-class FakeReader():
-    def __init__(self, data=b'', eof=False):
-        self.io = io.BytesIO(data)
-        self.eof = eof
-
-    async def readline(self):
-        return self.io.readline()
-
-    async def readexactly(self, n):
-        data = self.io.read(n)
-        if len(data) != n:
-            raise asyncio.IncompleteReadError(data, n)
-        return data
-
-    def at_eof(self):
-        return self.eof
 
 class TestEncoding(unittest.TestCase):
